script/pyproject.toml
[project]
name = "nlp-group"
version = "0.1.0"
description = ""
authors = [
    {name = "Reza Rahemtola",email = "reza.rahemtola@epitech.eu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
